requirements.txt
# 商城后台管理系统自动化测试依赖包

# 测试框架
pytest>=7.4.0
pytest-html>=3.2.0
pytest-xdist>=3.3.0
pytest-rerunfailures>=12.0
allure-pytest>=2.13.0

# HTTP请求
requests>=2.31.0

# 数据库
pymysql>=1.1.0
redis>=4.6.0
SQLAlchemy>=2.0.0

# UI自动化
selenium>=4.12.0
webdriver-manager>=4.0.0

# 配置文件解析
pyyaml>=6.0
python-dotenv>=1.0.0

# 数据处理
jsonpath>=0.82
jsonschema>=4.19.0

# 文件上传
paramiko>=3.3.0

# 日志
colorlog>=6.7.0

# 数据生成
faker>=19.0.0

# 工具库
python-dateutil>=2.8.2
pytz>=2023.3

# API文档测试（可选）
openpyxl>=3.1.0  # Excel操作
